pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "toolkit"
description = "supporting functions for dataset generartion and machine learning"
readme = "README.md"
version = "0.0.1"
authors = [
  { name = "Kirk Bonney", email = "klbonne@sandia.gov" },
]
maintainers = [
  { name = "Kirk Bonney", email = "klbonne@sandia.gov" },
]
requires-python = ">=3.8"

dependencies = [
    "numpy==1.26", 
    "scipy", 
    "pandas==2.2.0", 
    "matplotlib==3.8.2",
    "torch==2.2.1",
    "hmmlearn==0.3.0",
    "optuna==4.0.0",
    "h5py",
    "seaborn==0.13.1",
    "pysal",
    "esda",
    "pyyaml"
    ]

classifiers = [
  "License :: OSI Approved :: BSD License",
  "Programming Language :: Python :: 3 :: Only",
  "Programming Language :: Python :: 3.8",
  "Programming Language :: Python :: 3.9",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
]
